refactor(electron): reorganize directory structure
Reorganize src/lib/ into logical subdirectories:
- src/action/ for ActionExecutor and ActionScheduler
- src/registry/ for stateManagerRegistry
- src/subscription/ for SubscriptionManager
- src/thunk/ with subdirectories for Thunk, init, registration/, scheduling/

Update all import paths across source and test files to match new structure.
All tests passing (469/469) and validation suite clean.
