+++ packages/electron/src/bridge/ipc/IpcHandler.ts
@@ -4,10 +4,10 @@ import type { IpcMainEvent, IpcMainInvokeEvent } from 'electron';
 import { ipcMain } from 'electron';
 import { IpcChannel } from '../../constants.js';
 import { IpcCommunicationError } from '../../errors/index.js';
-import { thunkManager } from '../../lib/initThunkManager.js';
-import { getPartialState } from '../../lib/SubscriptionManager.js';
-import { Thunk as ThunkClass } from '../../lib/Thunk.js';
-import { ThunkRegistrationQueue } from '../../lib/ThunkRegistrationQueue.js';
+import { thunkManager } from '../../thunk/init.js';
+import { getPartialState } from '../../subscription/SubscriptionManager.js';
+import { Thunk as ThunkClass } from '../../thunk/Thunk.js';
+import { ThunkRegistrationQueue } from '../../thunk/registration/ThunkRegistrationQueue.js';
 import { actionQueue } from '../../main/actionQueue.js';
 import { logZubridgeError, serializeError } from '../../utils/errorHandling.js';
 import { sanitizeState } from '../../utils/serialization.js';
