+++ packages/electron/src/registry/stateManagerRegistry.ts
