+++ packages/electron/src/subscription/SubscriptionManager.ts
