+++ packages/electron/src/thunk/init.ts
@@ -1,7 +1,7 @@
 import { debug } from '@zubridge/core';
-import { initActionScheduler } from './ActionScheduler.js';
-import { initThunkManager as initManager } from './ThunkManager.js';
-import { ThunkScheduler } from './ThunkScheduler.js';
+import { initActionScheduler } from '../action/ActionScheduler.js';
+import { initThunkManager as initManager } from '../lib/ThunkManager.js';
+import { ThunkScheduler } from './scheduling/ThunkScheduler.js';
 
 // Initialize the ThunkManager immediately when this file is imported
 const scheduler = new ThunkScheduler();
